+++ CHANGES.md
@@ -6,9 +6,12 @@ v1.3.0 - Month DD, YYYY
 
 - Added new dithering algorithm to better support barcode printing with shaded
   content.
-- Added basic Brother printer support (Issue #15)
+- Added experimental Brother printer support (Issue #15)
 - Added basic TSPL printer support (Issue #54)
 - Added basic SEIKO printer support (Issue #58)
+- Added experimental Zebra PCL printer support.
+- Added support for configuration files in "/etc", "/usr/local/etc", or
+  "/Library/Application Support" (macOS).
 - Updated ZPL driver to report 'media-needed' reason when out of labels during a
   job.
 - Fixed copies support for ZPL printers (Issue #100)